backend/requirements.txt
flask-socketio
Flask
bcrypt
Flask-Cors
Flask-Jsonpify
Flask-RESTful
Flask-SQLAlchemy
pandas
Pillow
joblib
wikipedia
nltk
pymongo
pyspellchecker
requests
scapy
scikit-image
scikit-learn
scipy
seaborn
SpeechRecognition
# tensorflow
tflearn
PyJWT
python-dotenv